Artificial_intelligence/finalhomework/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
transformers>=4.0.0
pandas>=1.1.0
Pillow>=8.0
scikit-learn>=0.24.0
tqdm>=4.42.0